+++ packages/core/src/components/history-items/history-items.tsx
@@ -1,11 +1,11 @@
 import { Component, ComponentInterface, Host, h } from "@stencil/core";
 
 @Component({
-  tag: "dso-history-item-list",
-  styleUrl: "history-item-list.scss",
+  tag: "dso-history-items",
+  styleUrl: "history-items.scss",
   shadow: true,
 })
-export class HistoryItemList implements ComponentInterface {
+export class HistoryItems implements ComponentInterface {
   render() {
     return (
       <Host>