+++ packages/dso-toolkit/src/components/history-item-list/index.ts
@@ -1,3 +1,0 @@
-export * from "./history-item-list.args.js";
-export * from "./history-item-list.models.js";
-export * from "./history-item-list.stories-of.js";