+++ packages/dso-toolkit/src/components/history-items/history-items.content.ts
@@ -1,4 +1,4 @@
-import { HistoryItem } from "./history-item-list.models";
+import { HistoryItem } from "./history-items.models";
 
 export const headings = ["Datum", "Gebeurtenis"];
 
@@ -6,7 +6,7 @@ export const historyItems: HistoryItem[] = [
   {
     date: "02-08-2025",
     statusMessage: "Dit is de status message",
-    type: "Besluit",
+    type: "besluit",
     title: "Dit is de title. Deze heeft alles. Ik ben een Besluit",
     href: "#",
     explanation: "Een explanation",
@@ -15,26 +15,26 @@ export const historyItems: HistoryItem[] = [
   {
     date: "03-08-2025",
     statusMessage: "Dit heeft alleen de status message en is een Waarschuwing",
-    type: "Waarschuwing",
+    type: "waarschuwing",
   },
   {
     date: "04-08-2025",
     statusMessage: "Dit is de status message van een Tijdelijk Regelingdeel",
-    type: "Tijdelijk Regelingdeel",
+    type: "tijdelijk-regelingdeel",
     title: "Dit is de title. Deze heeft geen warning en explanation",
     href: "#",
   },
   {
     date: "05-08-2025",
     statusMessage: "Dit is de status message",
-    type: "In Werking",
+    type: "in-werking",
     title: "Dit is de title. Deze his In Werking",
     href: "#",
   },
   {
     date: "06-08-2025",
     statusMessage: "Dit is de status message",
-    type: "Ontwerp",
+    type: "ontwerp",
     title: "Dit is de title. Ik Ontwerp",
     href: "#",
   },