+++ storybook/src/components/history-items/history-items.core-stories.ts
@@ -1,24 +1,24 @@
 import historyItemReadme from "@dso-toolkit/core/src/components/history-item-list/components/history-item/readme.md?raw";
 import readme from "@dso-toolkit/core/src/components/history-item-list/readme.md?raw";
 import type { Meta } from "@storybook/web-components-vite";
-import { HistoryItemListArgs, historyItemListMeta, historyItemListStories } from "dso-toolkit";
+import { HistoryItemsArgs, historyItemsMeta, historyItemsStories } from "dso-toolkit";
 
 import { templateContainer } from "../../templates";
 
-const meta: Meta<HistoryItemListArgs> = {
-  ...historyItemListMeta({ readme: `${readme}\n${historyItemReadme}` }),
+const meta: Meta<HistoryItemsArgs> = {
+  ...historyItemsMeta({ readme: `${readme}\n${historyItemReadme}` }),
   title: "Core/History Item List",
 };
 
 export default meta;
 
-const { Default } = historyItemListStories({
+const { Default } = historyItemsStories({
   templateContainer,
   storyTemplates: (templates) => {
-    const { historyItemListTemplate } = templates;
+    const { historyItemsTemplate } = templates;
 
     return {
-      historyItemListTemplate,
+      historyItemsTemplate,
     };
   },
 });