+++ storybook/src/components/history-items/history-items.core-template.ts
@@ -1,20 +1,20 @@
 import { DsoHistoryItemCustomEvent } from "@dso-toolkit/core";
-import { HistoryItemClickEvent, HistoryItemList } from "dso-toolkit";
+import { HistoryItemClickEvent, HistoryItems } from "dso-toolkit";
 import { html, nothing } from "lit-html";
 
 import { ComponentImplementation } from "../../templates";
 
-export const coreHistoryItemList: ComponentImplementation<HistoryItemList> = {
-  component: "historyItemList",
+export const coreHistoryItems: ComponentImplementation<HistoryItems> = {
+  component: "historyItems",
   implementation: "core",
   template: () =>
-    function historyItemListTemplate({ headings, historyItems }) {
+    function historyItemsTemplate({ headings, historyItems }) {
       return html`<dso-history-item-list>
         ${headings.map((heading) => html`<span slot="heading">${heading}</span>`)}
         ${historyItems.map(
           ({ date, explanation, statusMessage, title, href, type, warning, dsoHistoryItemClick }) =>
             html`<dso-history-item
-              type=${type}
+              .type=${type}
               href=${href}
               @dsoHistoryItemClick=${(e: DsoHistoryItemCustomEvent<HistoryItemClickEvent>) => {
                 if (!e.detail.isModifiedEvent) {
@@ -24,7 +24,7 @@ export const coreHistoryItemList: ComponentImplementation<HistoryItemList> = {
                 dsoHistoryItemClick?.(e);
               }}
             >
-              ${html`<span slot="date">${date}</span>`} ${html`<span slot="status">${statusMessage}</span>`}
+              <span slot="date">${date}</span> <span slot="status">${statusMessage}</span>
               ${title ? html`<span slot="title">${title}</span>` : nothing}
               ${explanation ? html`<span slot="explanation">${explanation}</span>` : nothing}
               ${warning ? html`<span slot="warning">${warning}</span>` : nothing}