+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## Next
 
 ### Added
-* History Item: Nieuw component ([#3192](https://github.com/dso-toolkit/dso-toolkit/issues/3192))
+* History Items: Nieuwe component ([#3192](https://github.com/dso-toolkit/dso-toolkit/issues/3192))
 
 ### Changed
 * **BREAKING** Breadcrumbs: Breadcrumb met "s" ([#3210](https://github.com/dso-toolkit/dso-toolkit/issues/3210))